requirements.txt
glfw==2.5.6
imgui==2.0.0
numpy==1.24.2
pycairo==1.23.0
PyOpenGL==3.1.6
